Moved the asynctask to its own class.
